scheduled scraping functions as shells
